requirements.txt
discord.py>=2.0.0
requests>=2.28.0
types-requests>=2.28.0
python-dotenv>=0.20.0
huggingface-hub>=0.16.4
rich>=13.3.3
ollama>=0.1.0
pendulum>=3.0.0  # Better datetime handling with timezone support
dateparser>=1.1.0
types-dateparser>=1.1.0
sqlalchemy>=2.0.0

# Core packages for local vector search
llama-index-core>=0.10.0  # Base package with core functionality
llama-index-embeddings-ollama>=0.1.0  # Ollama embeddings
llama-index-llms-ollama>=0.1.0  # Ollama LLM
llama-index-vector-stores-chroma>=0.1.0  # ChromaDB vector store
chromadb>=0.4.22  # Local vector database

# Dependencies for lorekeeper
sentence-transformers>=2.6.0  # For embeddings
pymongo>=4.6.0  # MongoDB client
qdrant-client>=1.8.0  # Qdrant vector database

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
